+++ server/controller/UserController.js
@@ -1,11 +1,14 @@
 const Model = require('../db/models');
+const Utils = require('../../utils');
 const bcrypt = require('bcrypt');
 require('dotenv').config();
 
 const getAllUsers = (req, res) => {
-  Model.User.findAll()
+  Model.User.findAll({
+    attributes: { exclude: ['password'] },
+  })
     .then((result) => {
-      if (result.length < 1) {
+      if (result.length === 0) {
         return res
           .status(404)
           .json({ code: 404, message: "Don't have any user created yet!" });
@@ -22,9 +25,10 @@ const getUserById = (req, res, next) => {
     where: {
       id: req.params.id,
     },
+    attributes: { exclude: ['password'] },
   })
     .then((result) => {
-      if (result.length > 1) {
+      if (result !== null) {
         res.status(200).json(result);
       }
       return res.status(404).json({ code: 404, message: 'User not found!' });
@@ -37,32 +41,22 @@ const getUserById = (req, res, next) => {
 const postUser = (req, res) => {
   const { name, email, role, restaurant } = req.body;
   const password = req.body.password;
-  if (req.body.name == '') {
-    res.status(400).send({ code: '400', message: 'Name can not be empty' });
-  } else if (req.body.email == '') {
-    res.status(400).send({ code: '400', message: 'Email can not be empty' });
-  } else if (req.body.password == '') {
-    res.status(400).send({ code: '400', message: 'Password can not be empty' });
-  } else if (req.body.role == '') {
-    res.status(400).send({ code: '400', message: 'Role can not be empty' });
-  } else if (req.body.restaurant == '') {
-    res
-      .status(400)
-      .send({ code: '400', message: 'Restaurant can not be empty' });
-    // }else if( ) {
-    //   Model.User.findOne({
-    //     where: {
-    //       email: req.body.email
-    //     },
-    //   })
-    //     .then((result) => {
-    //       if (result.length > 1) {
-    //         res.status(200).json(result);
-    //       }
-    //       return res.status(404).json({ code: 404, message: 'User not found!' });
-    //     })
-  } else {
-    Model.User.create({
+  Utils.validationEmptyorNull(req.body.name, res, 'Name');
+  Utils.validationEmptyorNull(req.body.email, res, 'Email');
+  Utils.validationEmptyorNull(req.body.password, res, 'Password');
+  Utils.validationEmptyorNull(req.body.role, res, 'Role');
+  Utils.validationEmptyorNull(req.body.restaurant, res, 'Restaurant');
+  Model.User.findOne({
+    where: {
+      email: req.body.email,
+    },
+  }).then((result) => {
+    if (result !== null) {
+      return res
+        .status(403)
+        .json({ code: 403, message: 'Email already in use!' });
+    }
+    return Model.User.create({
       name,
       email,
       password: bcrypt.hashSync(password, 10),
@@ -79,9 +73,9 @@ const postUser = (req, res) => {
         });
       })
       .catch((err) => {
-        return res.status(400).json({ code: 500, message: err.message });
+        return res.status(400).json({ code: 400, message: err.message });
       });
-  }
+  });
 };
 
 const putUser = (req, res) => {
@@ -95,24 +89,33 @@ const putUser = (req, res) => {
     }
   )
     .then(() => {
-      res.status(200).send('Usuário alterado com sucesso');
+      getUserById(req, res);
     })
     .catch((err) => {
-      return res.status(400).json({ code: 500, message: err.message });
+      return res.status(400).json({ code: 400, message: err.message });
     });
 };
 
 const deleteUser = (req, res) => {
-  Model.User.destroy({
+  Model.User.findOne({
     where: {
       id: req.params.id,
     },
+    attributes: { exclude: ['password'] },
   })
-    .then(() => {
-      res.status(200).send('Usuário excluído com sucesso');
+    .then((result) => {
+      if (result !== null) {
+        res.status(200).json(result);
+        Model.User.destroy({
+          where: {
+            id: req.params.id,
+          },
+        });
+      }
+      return res.status(404).json({ code: 404, message: 'User not found!' });
     })
     .catch((err) => {
-      return res.status(400).json({ code: 500, message: err.message });
+      return res.status(400).json({ code: 400, message: err.message });
     });
 };
